PIMCORE-901
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@391 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/modules/admin/controllers/AssetController.php
@@ -226,8 +226,10 @@ class Admin_AssetController extends Pimcore_Controller_Action_Admin {
                         $found = true;
                         if (!$node->permissionSet) {
                             //reset permission by deleting it
-                            $assetPermission->delete();
-                            $permissions = $asset->getPermissions();
+                            if($assetPermission->getCid() == $asset->getId()){
+                                $assetPermission->delete();
+                                $permissions = $asset->getPermissions();
+                            }
                             break;
 
                         } else {
